requirements.txt
fastmcp==2.14.5
msal>=1.31.0
azure-search-documents>=11.7.0b2
azure-core>=1.32.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
click>=8.1.0
requests>=2.31.0
agent-framework-core==1.0.0b260116
azure-identity==1.25.1
agent-identity-python-sdk>=0.1.2
